backend/requirements.txt
flask==2.3.3
flask-cors==4.0.0
python-dotenv==1.0.0
gunicorn==21.2.0
requests==2.31.0
beautifulsoup4==4.12.2
google-search-results==2.4.2
pandas==1.5.3
matplotlib==3.7.2
selenium==4.15.2
webdriver-manager==4.0.1
nltk==3.8.1
# 移除google-generativeai依赖，不再使用Gemini API
